refactor: ref Podcasts
