--- src/view/containers/podcast/components/PodcastDetail/index.js
@@ -1,1 +1,0 @@
-export { default } from './PodcastDetail';
