+++ src/view/containers/podcast/components/PodcastEpisodesTable/PodcastEpisodesTable.tsx
@@ -9,18 +9,12 @@ const PodcastEpisodesTable = ({ podcasts }) => {
       <table className={styles.table}>
         <thead className={styles.table__header}>
           <tr>
-            <th
-              className={styles.table__title}
-              itemScope="col"
-              align="left"
-              id="col-title"
-            >
+            <th className={styles.table__title} align="left" id="col-title">
               Title
             </th>
             <th
               className={styles.table__title}
               align="right"
-              itemScope="col"
               style={{ textAlign: 'left' }}
               id="col-date"
             >
@@ -29,7 +23,6 @@ const PodcastEpisodesTable = ({ podcasts }) => {
             <th
               className={styles.table__title}
               align="right"
-              itemScope="col"
               style={{ textAlign: 'left' }}
               id="col-duration"
             >
