--- src/view/containers/podcast/components/PodcastEpisodesTable/index.js
@@ -1,1 +1,0 @@
-export { default } from './PodcastEpisodesTable';
