--- src/view/containers/podcast/components/PodcastList/index.js
@@ -1,1 +1,0 @@
-export { default } from './PodcastList';
