--- src/view/containers/podcast/components/Results/index.js
@@ -1,1 +1,0 @@
-export { default } from './Results';
