--- src/view/containers/podcast/index.js
@@ -1,3 +1,0 @@
-export { default } from './Podcast';
-
-export { loader } from './Podcast';
