+++ src/view/containers/routes/Routes.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import ErrorPage from './Error';
 import Main from '../main/Main';
-import Podcast, { loader as podcastLoader } from '../podcast';
+import Podcast, { loader as podcastLoader } from '../podcast/Podcast';
 import Episode from '../episode/Episode';
 import Search, { loader as searchLoader } from '../search';
 import PodcastList from '../podcast/components/PodcastList/PodcastList';
