+++ src/view/containers/routes/__tests__/Routes.test.js
@@ -6,7 +6,7 @@ jest.mock('../Error', () => jest.fn());
 jest.mock('../../main/Main', () => jest.fn());
 jest.mock('../../episode/Episode', () => jest.fn());
 jest.mock('../../podcast/components/PodcastList/PodcastList', () => jest.fn());
-jest.mock('../../podcast', () => {
+jest.mock('../../podcast/Podcast', () => {
   return {
     __esModules: true,
     default: jest.fn(),
